move peers api to node_info, and don't return it by default in the client.node_info() api
unblock users of client api to fetch chain_info when p2p is disabled.
